feat(lambda): move to a single github client with conditional requests
We're currently creating fresh clients for each lambda invocation. This
is not entirely optimal, because we'll potentially request the same data
from the API multiple times.

Move instead to a single central client. This is created at the module
scope and then passed down into the handlers. The client also uses
`toad-cache` to cache previously-seen responses and send `If-None-Match`
and `If-Modified-Since` headers to the API to allow it to avoid sending
us back data we already have in our hands. When this happens we get a
"304 Not Modified" response, we use our copy, and the call doesn't
consume the rate limit.